Lots of misc improvements.
- GLFramebufferObject: added bindExternalFBO() to bind an externally
  allocated FBOs such as from QOpenGLWidget.
- GLTexture
  - Moved targets into its own enum class.
  - Got rid of GLtexture::get() functions: they belong in subclasses.
  - Added mipmap support to GLTexture2D.
- Removed redundant getOrtho() / setOrtho() functions in OrthographicCamera.
- Renamed some functions in PerspectiveCamera.
- ArrayUtils
  - saveTXT() now uses Array2DView / Array3DView instead of Array2D /
  Array3D.
  - Added componentView() to extract components out of an ArrayView<complex data type>
- BoxUtils learned some extra primitives.
- RectnagleUtils learned some useful methods that are well documented.
- Removed functions from D3D11RectangleUtils / D3D11Utils_Box: they're
  now in geometry/BoxUtils / RectangleUtils.
- Removed GeometryUtils::pointInBox(): box has that as a function.
- Arithmetic finally has a log2.
- Finally standardized rectangle/box conventions.

+++ GL/include/GLBufferObject.h
@@ -126,7 +126,7 @@ Array1DView< T > GLBufferObject::mapAs( GLBufferObject::Access access )
 {
     return Array1DView< T >(
         glMapNamedBufferEXT( m_id, static_cast< GLbitfield >( access ) ),
-        m_nBytes / sizeof( T ) );
+        static_cast< int >( m_nBytes / sizeof( T ) ) );
 }
 
 template< typename T >
